Addresses #100: move ZoomManager to draw 2d
This is an attempt to move the common elements of ZoomManager and
ZoomListener class to draw2d as an abstract base class. This includes
already a first clean-up of ZoomManager class as well as the usage of
the listeners.
