+++ org.eclipse.draw2d/META-INF/MANIFEST.MF
@@ -2,17 +2,18 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: %Plugin.name
 Bundle-SymbolicName: org.eclipse.draw2d
-Bundle-Version: 3.11.200.qualifier
+Bundle-Version: 3.12.0.qualifier
 Bundle-Vendor: %Plugin.providerName
 Bundle-Localization: plugin
 Export-Package: org.eclipse.draw2d,
- org.eclipse.draw2d.graph,
  org.eclipse.draw2d.geometry,
+ org.eclipse.draw2d.graph,
  org.eclipse.draw2d.images,
  org.eclipse.draw2d.internal;x-internal:=true,
  org.eclipse.draw2d.parts,
  org.eclipse.draw2d.text,
- org.eclipse.draw2d.widgets
+ org.eclipse.draw2d.widgets,
+ org.eclipse.draw2d.zoom
 Import-Package: com.ibm.icu.text
 Require-Bundle: org.eclipse.swt;visibility:=reexport;bundle-version="[3.4.0,4.0.0)"
 Bundle-ActivationPolicy: lazy
